refactor: reorganize deploy/revert/verify operations into clear directory structure
- Create operations/ directory with modules/, projects/ subdirectories
- Use consistent deploy.ts/revert.ts/verify.ts naming throughout
- Move module-level operations to operations/modules/
- Move project-level operations to operations/projects/
- Move main routing logic to operations/index.ts
- Update all import paths and exports
- Maintain backward compatibility with existing function names
