+++ packages/core/src/operations/modules/revert.ts
@@ -1,17 +1,17 @@
 import { join } from 'path';
 import { existsSync } from 'fs';
-import { LaunchQLMigrate } from './client';
-import { MigrateConfig } from './types';
+import { LaunchQLMigrate } from '../../migrate/client';
+import { MigrateConfig } from '../../migrate/types';
 import { Logger } from '@launchql/logger';
 
-const log = new Logger('migrate-deploy');
+const log = new Logger('migrate-revert');
 
 /**
- * Deploy command that mimics sqitch deploy behavior
- * This is designed to be a drop-in replacement for spawn('sqitch', ['deploy', 'db:pg:database'])
+ * Revert command that mimics sqitch revert behavior
+ * This is designed to be a drop-in replacement for spawn('sqitch', ['revert', 'db:pg:database'])
  */
-export async function deployModule(
-  config: Partial<MigrateConfig>,
+export async function revertModule(
+  config: any,
   database: string,
   cwd: string,
   options?: {
@@ -25,19 +25,16 @@ export async function deployModule(
     throw new Error(`No launchql.plan found in ${cwd}`);
   }
   
-  // Provide defaults for missing config values
   const fullConfig: MigrateConfig = {
-    host: config.host,
-    port: config.port,
-    user: config.user,
-    password: config.password,
-    database: config.database
+    pg: config,
+    database: database,
+    modulePath: cwd
   };
   
   const client = new LaunchQLMigrate(fullConfig);
   
   try {
-    const result = await client.deploy({
+    const result = await client.revert({
       project: '', // Will be read from plan file
       targetDatabase: database,
       planPath,
@@ -46,12 +43,12 @@ export async function deployModule(
     });
     
     if (result.failed) {
-      throw new Error(`Deployment failed at change: ${result.failed}`);
+      throw new Error(`Revert failed at change: ${result.failed}`);
     }
     
-    log.info(`Deployed ${result.deployed.length} changes`);
+    log.info(`Reverted ${result.reverted.length} changes`);
     if (result.skipped.length > 0) {
-      log.info(`Skipped ${result.skipped.length} already deployed changes`);
+      log.info(`Skipped ${result.skipped.length} not deployed changes`);
     }
   } finally {
     // Pool is managed by PgPoolCacheManager, no need to close
