+++ packages/core/src/operations/projects/deploy.ts
@@ -4,10 +4,10 @@ import { errors, LaunchQLOptions } from '@launchql/types';
 import { PgConfig } from 'pg-env';
 import { Logger } from '@launchql/logger';
 import { getPgPool } from 'pg-cache';
-import { deployModule } from '../migrate/deploy-module';
-import { LaunchQLProject } from '../class/launchql';
-import { packageModule } from '../package';
-import { runSqitch } from '../utils/sqitch-wrapper';
+import { deployModule } from '../modules/deploy';
+import { LaunchQLProject } from '../../class/launchql';
+import { packageModule } from '../../package';
+import { runSqitch } from '../../utils/sqitch-wrapper';
 
 interface Extensions {
   resolved: string[];
